Combine, Shooting Bug

+++ Assets/Scripts/HealthBar.cs
@@ -6,7 +6,7 @@ public class HealthBar : MonoBehaviour {
     public Gradient gradient;
     public Image fill;
 
-    public void SetHealth(int health) {
+    public void SetHealth(float health) {
         slider.value = health;
 
         fill.color = gradient.Evaluate(slider.normalizedValue);
